+++ tests/testthat/test_simulation.R
@@ -6,10 +6,9 @@ if (requireNamespace("igraph", quietly = TRUE)) {
     data(cascades)
     test_that("Simulation function works.", {
         from_netinf <- netinf(cascades, params = 1, trans_mod = "exponential",
-                              n_edges = 5, quiet = TRUE)
-        out <- simulate_cascades(from_netinf, nsim = 100, seed = 123, max_time = 10, 
-                                 lambda = 1, beta = 0.5, epsilon = 10e-9, 
-                                 model = "exponential")
+                              n_edges = 5, quiet = TRUE, max_iter = 1)
+        set.seed(123)
+        out <- simulate_cascades(from_netinf, nsim = 100, max_time = 10)
         casc <- as_cascade_long(out)
         rec <- netinf(casc, params = 1, trans_mod = "exponential", n_edges = 5,
                       quiet = TRUE)
@@ -26,9 +25,7 @@ if (requireNamespace("igraph", quietly = TRUE)) {
         partial_cascade$cascade_times <- cascades$cascade_times[10]
         from_netinf <- netinf(cascades, params = 1, trans_mod = "exponential",
                               n_edges = 100, quiet = TRUE)   
-        out <- simulate_cascades(from_netinf, nsim = 100, seed = 123, max_time = 10, 
-                                 lambda = 1, beta = 0.5, epsilon = 10e-9, 
-                                 model = "exponential", 
+        out <- simulate_cascades(from_netinf, nsim = 100, max_time = 10, 
                                  partial_cascade = partial_cascade)
         expect_equal(length(unique(out[out$node_name == "22", 2])), 1)
         expect_equal(length(unique(out[out$node_name == "28", 2])), 1)